# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AllTransactionRow": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/CardResponse"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionMatch"
          }
        },
        "required": [
          "transaction",
          "card"
        ],
        "title": "AllTransactionRow",
        "type": "object"
      },
      "AllTransactionsResponse": {
        "properties": {
          "bank_count": {
            "title": "Bank Count",
            "type": "integer"
          },
          "cache_as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache As Of"
          },
          "card_count": {
            "title": "Card Count",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/AllTransactionRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "total_matches": {
            "title": "Total Matches",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "total_matches",
          "card_count",
          "bank_count",
          "rows",
          "next_cursor",
          "has_more",
          "cache_as_of"
        ],
        "title": "AllTransactionsResponse",
        "type": "object"
      },
...
      "CardResponse": {
        "properties": {
          "bank": {
            "enum": [
              "capital-one",
              "chase",
              "citi",
              "wells-fargo"
            ],
            "title": "Bank",
            "type": "string"
          },
          "bank_display_name": {
            "title": "Bank Display Name",
            "type": "string"
          },
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_successful_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Successful Sync At"
          },
          "mask": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "official_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Official Name"
          },
          "state": {
            "default": "ready",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "connection_id",
          "bank",
          "bank_display_name",
          "name",
          "official_name",
          "mask"
        ],
        "title": "CardResponse",
        "type": "object"
      },
      "CardTransactionGroup": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/CardResponse"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "match_count": {
            "title": "Match Count",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionMatch"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "card",
          "transactions",
          "match_count",
          "next_cursor",
          "has_more"
        ],
        "title": "CardTransactionGroup",
        "type": "object"
      },
...
      "CreateLinkTokenRequest": {
        "properties": {
          "bank": {
            "enum": [
              "capital-one",
              "chase",
              "citi",
              "wells-fargo"
            ],
            "title": "Bank",
            "type": "string"
          },
          "confirm_trial_slot": {
            "default": false,
            "title": "Confirm Trial Slot",
            "type": "boolean"
          }
        },
        "required": [
          "bank"
        ],
        "title": "CreateLinkTokenRequest",
        "type": "object"
      },
...
      "CreateTransactionRefreshResponse": {
        "properties": {
          "coalesced": {
            "title": "Coalesced",
            "type": "boolean"
          },
          "refresh": {
            "$ref": "#/components/schemas/TransactionRefreshResponse"
          }
        },
        "required": [
          "coalesced",
          "refresh"
        ],
        "title": "CreateTransactionRefreshResponse",
        "type": "object"
      },
...
      "ExchangePublicTokenRequest": {
        "properties": {
          "bank": {
            "enum": [
              "capital-one",
              "chase",
              "citi",
              "wells-fargo"
            ],
            "title": "Bank",
            "type": "string"
          },
          "institution_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Institution Id",
            "type": "string"
          },
          "institution_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Institution Name",
            "type": "string"
          },
          "public_token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Public Token",
            "type": "string"
          }
        },
        "required": [
          "bank",
          "public_token",
          "institution_id",
          "institution_name"
        ],
        "title": "ExchangePublicTokenRequest",
        "type": "object"
      },
      "ExchangeResponse": {
        "properties": {
          "bank": {
            "enum": [
              "capital-one",
              "chase",
              "citi",
              "wells-fargo"
            ],
            "title": "Bank",
            "type": "string"
          },
          "card_count": {
            "title": "Card Count",
            "type": "integer"
          },
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          },
          "institution_name": {
            "title": "Institution Name",
            "type": "string"
          },
          "sync_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Job Id"
          }
        },
        "required": [
          "connection_id",
          "bank",
          "institution_name",
          "card_count",
          "sync_job_id"
        ],
        "title": "ExchangeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LinkTokenResponse": {
        "properties": {
          "bank": {
            "title": "Bank",
            "type": "string"
          },
          "consumes_trial_slot": {
            "title": "Consumes Trial Slot",
            "type": "boolean"
          },
          "link_token": {
            "title": "Link Token",
            "type": "string"
          },
          "mode": {
            "enum": [
              "new",
              "update"
            ],
            "title": "Mode",
            "type": "string"
          },
          "production_item_count": {
            "title": "Production Item Count",
            "type": "integer"
          },
          "production_item_limit": {
            "title": "Production Item Limit",
            "type": "integer"
          }
        },
        "required": [
          "link_token",
          "bank",
          "mode",
          "consumes_trial_slot",
          "production_item_count",
          "production_item_limit"
        ],
        "title": "LinkTokenResponse",
        "type": "object"
      },
...
      "TransactionMatch": {
        "properties": {
          "amount_cents": {
            "title": "Amount Cents",
            "type": "integer"
          },
          "authorized_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized Date"
          },
          "card_id": {
            "title": "Card Id",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "currency_code": {
            "title": "Currency Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "merchant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Name"
          },
          "original_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Description"
          },
          "pending": {
            "title": "Pending",
            "type": "boolean"
          },
          "posted_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted Date"
          }
        },
        "required": [
          "id",
          "card_id",
          "merchant_name",
          "description",
          "original_description",
          "category",
          "amount_cents",
          "currency_code",
          "authorized_date",
          "posted_date",
          "pending"
        ],
        "title": "TransactionMatch",
        "type": "object"
      },
      "TransactionRefreshResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "partial",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/TransactionRefreshSummaryResponse"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/TransactionRefreshTargetResponse"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "id",
          "state",
          "created_at",
          "started_at",
          "finished_at",
          "expires_at",
          "summary",
          "targets"
        ],
        "title": "TransactionRefreshResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Transaction Aggregator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cards/{card_id}/transactions": {
      "get": {
        "operationId": "card_transactions_api_cards__card_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "title": "Card Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardTransactionGroup"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Card Transactions",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/connections/exchange": {
      "post": {
        "operationId": "exchange_public_token_api_connections_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangePublicTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Public Token",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/link-token": {
      "post": {
        "operationId": "create_link_token_api_connections_link_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Link Token",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/{connection_id}": {
      "delete": {
        "operationId": "disconnect_api_connections__connection_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disconnect",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/api/connections/{connection_id}/update-token": {
      "post": {
        "operationId": "create_update_link_token_api_connections__connection_id__update_token_post",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Update Link Token",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/api/transaction-refreshes": {
      "post": {
        "operationId": "create_transaction_refresh_api_transaction_refreshes_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[\\x21-\\x7e]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTransactionRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Transaction Refresh",
        "tags": [
          "transaction-refreshes"
        ]
      }
    },
    "/api/transaction-refreshes/active": {
      "get": {
        "operationId": "active_transaction_refresh_api_transaction_refreshes_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No active transaction refresh"
          }
        },
        "summary": "Active Transaction Refresh",
        "tags": [
          "transaction-refreshes"
        ]
      }
    },
    "/api/transaction-refreshes/{refresh_id}": {
      "get": {
        "operationId": "get_transaction_refresh_api_transaction_refreshes__refresh_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "refresh_id",
            "required": true,
            "schema": {
              "title": "Refresh Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Transaction Refresh",
        "tags": [
          "transaction-refreshes"
        ]
      }
    },
    "/api/transactions": {
      "get": {
        "operationId": "all_transactions_api_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllTransactionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "All Transactions",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}